instrumentations/microsoft-foundry/pyproject.toml
[project]
name = "llm-tracekit-microsoft-foundry"
version = "1.0.0"
description = "OpenTelemetry instrumentation for Microsoft Foundry (azure-ai-projects)."
readme = "README.md"
authors = [
    { name="Coralogix Ltd.", email="info@coralogix.com" },
]
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.10"
dependencies = [
    "llm-tracekit-core>=1.0.0",
    "opentelemetry-instrumentation>=0.60b0",
    "azure-ai-projects>=2.0.0",
    "wrapt>=1.0.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: OpenTelemetry :: Instrumentations",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

[project.urls]
homepage = "https://coralogix.com"
repository = "https://github.com/coralogix/llm-tracekit.git"

[project.entry-points.opentelemetry_instrumentor]
microsoft_foundry = "llm_tracekit.microsoft_foundry:MicrosoftFoundryInstrumentor"

[dependency-groups]
dev = [
    "assertpy>=1.1",
    "azure-identity>=1.15.0",
    "brotli>=1.1.0",
    "mypy>=1.14.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=1.0.0",
    "pytest-vcr>=1.0.2",
    "pyyaml>=6.0.2",
    "ruff>=0.9.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_tracekit"]

[tool.hatch.build.targets.sdist]
only-include = [
    "src/llm_tracekit",
]

[tool.uv.sources]
llm-tracekit-core = { workspace = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
